refactor: create new folder for fhe trait - move encryptable and decryptable from utils/fhe.rs to fhe/
