+++ src/lib.rs
@@ -5,7 +5,12 @@ pub mod configs {
 
 pub mod utils {
     pub mod base64;
-    pub mod fhe;
     pub mod file_ctl;
     pub mod json;
+}
+
+pub mod fhe {
+    pub mod computable;
+    pub mod decryptable;
+    pub mod encryptable;
 }
