+++ src/main.rs
@@ -1,4 +1,5 @@
 mod configs;
+mod fhe;
 mod utils;
 use std::fs::File;
 use std::io::Result;
