+++ src/utils/mod.rs
@@ -1,4 +1,3 @@
 pub mod base64;
-pub mod fhe;
 pub mod file_ctl;
 pub mod json;
