+++ tests/fhe.rs
@@ -1,7 +1,8 @@
 #[cfg(test)]
 mod fhe_tests {
     use fhe_http::configs::json::{FheJsonValue, NorJsonValue};
-    use fhe_http::utils::fhe::{Decryptable, Encryptable};
+    use fhe_http::fhe::decryptable::Decryptable;
+    use fhe_http::fhe::encryptable::Encryptable;
     use tfhe::{generate_keys, ConfigBuilder, Error};
 
     #[test]
